+++ src/lib/conf/.may_include
@@ -1,0 +1,2 @@
+orconfig.h
+lib/cc/*.h
