+++ src/lib/conf/include.am
@@ -1,0 +1,4 @@
+
+# ADD_C_FILE: INSERT HEADERS HERE.
+noinst_HEADERS +=				\
+	src/lib/conf/conftypes.h
